Update pace cloud sdk
