+++ App/Sources/AppDelegate.swift
@@ -14,21 +14,15 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
     ) -> Bool {
         PACECloudSDK.shared.setup(
             with: .init(
-                apiKey: Secrets.apiKey,
+                apiKey: <#Your API Key#>,
                 authenticationMode: .native,
-                environment: .sandbox,
-                customOIDConfiguration: .init(
-                    authorizationEndpoint: "\(Secrets.cloudURL)/auth/realms/pace/protocol/openid-connect/auth",
-                    tokenEndpoint: "\(Secrets.cloudURL)/auth/realms/pace/protocol/openid-connect/token",
-                    userEndpoint: "\(Secrets.cloudURL)/auth/realms/pace/protocol/openid-connect/userinfo",
-                    clientId: Secrets.clientID,
-                    redirectUri: Secrets.redirectURI
-                ),
+                environment: sdkEnvironment,
+                customOIDConfiguration: nil,
                 isRedirectSchemeCheckEnabled: true,
                 domainACL: ["pace.cloud"],
                 allowedLowAccuracy: nil,
                 speedThresholdInKmPerHour: nil,
-                geoAppsScope: "pace-drive-ios"
+                geoAppsScope: "pace-drive-ios-min"
             )
         )
