+++ App/Sources/Coordinators/PaymentHistory/PaymentHistoryCoordinator.swift
@@ -3,16 +3,15 @@ import JLCoordinator
 import PACECloudSDK
 import UIKit
 
-final class PaymentHistoryCoordinator: Coordinator {
+final class PaymentHistoryCoordinator: Coordinator, AppKitDelegate {
     typealias Callback = () -> Void
 
     private let callback: Callback?
+    private var appKitDelegate: AppKitDelegate? // swiftlint:disable:this weak_delegate
     private lazy var viewController: AppViewController = AppKit.appViewController(
         presetUrl: .transactions,
+        isModalInPresentation: false,
         completion: callback
-    )
-    private lazy var receiptCoordinator: PaymentReceiptCoordinator = .init(
-        presenter: ModalPresenter(presentingViewController: viewController)
     )
 
     init(presenter: Presenter, callback: Callback? = nil) {
@@ -23,6 +22,20 @@ final class PaymentHistoryCoordinator: Coordinator {
 
     override func start() {
         presenter.present(viewController, animated: true)
-        receiptCoordinator.start()
+        appKitDelegate = AppKit.delegate
+        AppKit.delegate = self
+    }
+
+    override func stop() {
+        AppKit.delegate = appKitDelegate
+        appKitDelegate = nil
+    }
+
+    func didFail(with error: AppKit.AppError) {
+        // Add error handling
+    }
+
+    func didReceiveAppDrawers(_ appDrawers: [AppKit.AppDrawer], _ appDatas: [AppKit.AppData]) {
+        // Add handling for app drawers
     }
 }
