+++ ConnectedFueling.xcodeproj/project.xcworkspace/xcshareddata/swiftpm/Package.resolved
@@ -24,8 +24,8 @@
         "repositoryURL": "https://github.com/pace/cloud-sdk-ios",
         "state": {
           "branch": null,
-          "revision": "921fed4e7ac3a203872fca4f7f79d4ad3f23e023",
-          "version": "8.0.0"
+          "revision": "7169827f8af6e91c3e989f2751acf17b759f2572",
+          "version": "9.0.0"
         }
       },
       {
